requirements.txt
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
colorclass==2.2.2
Cython==0.29.32
docopt==0.6.2
et-xmlfile==1.1.0
idna==3.4
joblib==1.2.0
libretranslatepy==2.1.1
lxml==4.9.1
nltk==3.7
numpy==1.23.5
oauthlib==3.2.2
openpyxl==3.0.10
packaging==21.3
pandas==1.5.1
pip-upgrader==1.4.15
pycld3==0.22
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.6
regex==2022.10.31
requests==2.28.1
requests-oauthlib==1.3.1
six==1.16.0
terminaltables==3.1.10
textblob==0.17.1
tqdm==4.64.1
translate==3.6.1
tweepy==4.12.1
urllib3==1.26.12
python-dotenv==0.21.0